requirements.txt
scikit-learn
matplotlib
numpy
pandas
tensorflow-cpu
keras
flask
pillow
seaborn
